requirements.txt
torch>=2.2.0
torchvision>=0.17.0
wandb
pillow
numpy
opencv-python
tqdm
einops
matplotlib